requirements.txt
apache-airflow
apache-airflow-providers-apache-kafka
google-cloud-bigquery
redis
pyflink
